refactor: Move bitmask utilities into src
It's fine keeping them in another file, but extracting them to include
is some form of premature optimization. By the time I feel confident in
stamping this project as done, I will need to move a lot more stuff in
there, or nothing at all. Let's go with the flow for now.