fix: Initialize empty display before main
macOS (or Raylib + macOS; Linux is unaffected) does not initialize the
application window if nothing gets drawn to the screen, so force a
single, empty buffer drawing action before starting the main loop.

This could have been done inside platform_init, but, in retrospect, it
is probably a good idea to do this in main to ensure that any display
backend is always initialized and cleared on startup.