fix: Correctly handle flag register
Previously, the flag register was only being set and never cleared. This
commit fixes it to always change value when an instruction that may
affect the flag register is executed. This also comes with a slight
refactoring of the arithmetic handling to support edge cases discovered
in test roms where one of the input registers may be the flag register
itself, meaning x, y, and f need to be decoupled from eachother to
prevent invalid assignments.

To facilitate this transition, the unit tests for shifting have been
split into separate tests to test behavior both with and without the
flag register being set. Other unit tests are unaffected as I had
apparently not written them, and so will leave them to a later MR when
some larger scale general cleanup is done.