fix: Prevent freeze after test ROMs receive input
Timendus' CHIP-8 test suite's menus use a function that waits for all
keys to be released before advancing to the next screen. Raylib only
polls inputs once per frame - during the `EndDrawing` function, which we
only call if the frame buffer was touched, meaning, in this specific
scenario, the emulator freezes in an infinite loop once a valid key was
registered. Not sure how common such event handling is in real ROMs, but
it sure was an issue that took a while to figure out